uint32_t -> uintptr_t for coding style

+++ kernel/task.c
@@ -117,7 +117,7 @@ int task_create()
     panic("Not enough memory for per process page directory!\n");
 
   /* Setup User Stack */
-	for(uint32_t i=USTACKTOP-USR_STACK_SIZE;i<USTACKTOP;i+=PGSIZE)
+	for(uintptr_t i=USTACKTOP-USR_STACK_SIZE;i<USTACKTOP;i+=PGSIZE)
 	{
 		struct PageInfo *pp = NULL;
 		if((pp = page_alloc(ALLOC_ZERO))==NULL)
@@ -164,7 +164,7 @@ int task_create()
 static void task_free(int pid)
 {
 	lcr3(PADDR(kern_pgdir));
-	for(uint32_t i=USTACKTOP-USR_STACK_SIZE;i<USTACKTOP;i+=PGSIZE)
+	for(uintptr_t i=USTACKTOP-USR_STACK_SIZE;i<USTACKTOP;i+=PGSIZE)
 		page_remove(tasks[pid].pgdir,(void*)i);
 	ptable_remove(tasks[pid].pgdir);
 	pgdir_remove(tasks[pid].pgdir);
